requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Deep Learning
torch>=2.0.0
einops>=0.7.0
transformers>=4.35.0

# Diffusion-TS (install from source)
# git+https://github.com/Y-debug-sys/Diffusion-TS.git

# Time Series Foundation Models
uni2ts>=0.1.0  # Moirai

# Visualization
plotly>=5.18.0
streamlit>=1.29.0

# Data
pyarrow>=14.0.0
h5py>=3.10.0

# Utilities
pyyaml>=6.0.1
tqdm>=4.66.0
loguru>=0.7.2

# Optional: Traditional IDS
pyshark>=0.6  # For pcap parsing (optional)
